Done
I added code to search the directory and open each formatted file.  It
then searches for a given recipe file based on OTP burns.

I had to screw around with the path naming though.  For some reason it
was having issues with '\' vs '/'.  It would combine them ??? and then
not load.